refactor: 좌석 선택 화면으로 이동하도록 수정
